PAE/requirements.txt
# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy>=2.0
pymysql>=1.1
cryptography>=42.0        # pymysql optional dep for caching_sha2_password

# ── Configuration ─────────────────────────────────────────────────────────────
pydantic>=2.0
pydantic-settings>=2.0
python-dotenv>=1.0

# ── HTTP / Scraping ───────────────────────────────────────────────────────────
requests>=2.31
feedparser>=6.0
beautifulsoup4>=4.12
lxml>=5.0

# ── LLMs ─────────────────────────────────────────────────────────────────────
anthropic>=0.25

# ── Notifications ─────────────────────────────────────────────────────────────
python-telegram-bot>=21.0

# ── Scheduling ────────────────────────────────────────────────────────────────
APScheduler>=3.10

# ── Trading ───────────────────────────────────────────────────────────────────
alpaca-py>=0.18           # official Alpaca SDK (replaces alpaca-trade-api)

# ── Data ──────────────────────────────────────────────────────────────────────
pandas>=2.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=8.0
pytest-asyncio>=0.23
